refactor: Rename warnings.py to fix examples workflow (#3734)
* ci: Fix examples workflow

* chore: adding changelog file 3732.maintenance.md [dependabot-skip]

* ci: Fix examples workflow

* ci: Fix examples workflow

* ci: Fix examples workflow

* chore: adding changelog file 3734.miscellaneous.md [dependabot-skip]

* ci: Fix examples workflow

* ci: Fix examples workflow

* ci: Fix examples workflow

* ci: Fix examples workflow

* ci: Fix examples workflow

* ci: Fix examples workflow

* ci: Fix examples workflow

* ci: Fix examples workflow

* ci: Fix examples workflow

* ci: Fix examples workflow

---------
