+++ doc/changelog.d/3732.maintenance.md
@@ -1,0 +1,1 @@
+Fix examples workflow [skip tests]
