+++ doc/changelog.d/3734.miscellaneous.md
@@ -1,0 +1,1 @@
+Rename warnings.py to fix examples workflow
