+++ src/ansys/fluent/core/services/deprecated_field_data.py
@@ -26,6 +26,7 @@ from typing import Callable, Dict, List
 import warnings
 
 from ansys.api.fluent.v0 import field_data_pb2 as FieldDataProtoModule
+from ansys.fluent.core.pyfluent_warnings import PyFluentDeprecationWarning
 from ansys.fluent.core.services.field_data import (
     ChunkParser,
     FieldDataService,
@@ -42,7 +43,6 @@ from ansys.fluent.core.services.field_data import (
     override_help_text,
 )
 from ansys.fluent.core.utils.deprecate import deprecate_argument
-from ansys.fluent.core.warnings import PyFluentDeprecationWarning
 
 DEPRECATION_MSG = "'field_data_old' is deprecated. Use 'field_data' instead."
