+++ src/ansys/fluent/core/utils/deprecate.py
@@ -26,7 +26,7 @@ import functools
 import logging
 import warnings
 
-from ansys.fluent.core.warnings import PyFluentDeprecationWarning
+from ansys.fluent.core.pyfluent_warnings import PyFluentDeprecationWarning
 
 logger = logging.getLogger("pyfluent.general")
