+++ src/ansys/fluent/core/utils/file_transfer_service.py
@@ -29,9 +29,9 @@ import shutil
 from typing import Any, Protocol
 import warnings
 
+from ansys.fluent.core.pyfluent_warnings import PyFluentUserWarning
 from ansys.fluent.core.utils import get_user_data_dir
 from ansys.fluent.core.utils.deprecate import deprecate_argument
-from ansys.fluent.core.warnings import PyFluentUserWarning
 import ansys.platform.instancemanagement as pypim
 
 # Host path which is mounted to the file-transfer-service container
